refactor(build): isolate n2 execution and remove stale adapters
