+++ crates/moon/Cargo.toml
@@ -39,6 +39,7 @@ semver.workspace = true
 clap.workspace = true
 anyhow.workspace = true
 colored.workspace = true
+ariadne.workspace = true
 serde.workspace = true
 serde_json_lenient.workspace = true
 serde_json.workspace = true
@@ -57,7 +58,6 @@ indexmap.workspace = true
 petgraph.workspace = true
 rand.workspace = true
 tempfile.workspace = true
-unicode-width.workspace = true
 shlex.workspace = true
 derive_builder.workspace = true
 sha2.workspace = true
