+++ crates/moon/src/rr_build/mod.rs
@@ -32,13 +32,11 @@ use std::{
     io::Write,
     path::{Path, PathBuf},
     rc::Rc,
-    sync::{LazyLock, mpsc},
+    sync::LazyLock,
 };
 
 use anyhow::Context;
-use clap::ValueEnum;
 use indexmap::IndexMap;
-use moonbuild::entry::{N2RunStats, ResultCatcher, create_progress_console};
 use moonbuild_rupes_recta::{
     CompileConfig, ResolveConfig, ResolveOutput,
     build_lower::WarningCondition,
@@ -50,7 +48,7 @@ use moonbuild_rupes_recta::{
         BackendConfig, DebugInfoRequest, DebugSymbols, ENV_MOONBIT_NEW_NATIVE, NativeTarget,
         OperatingSystem, PackageId, TargetKind,
     },
-    target_layout::{ArtifactPathResolver, GENERATED_TEST_DRIVER_PREFIX, TargetLayout},
+    target_layout::{ArtifactPathResolver, TargetLayout},
 };
 use moonutil::{
     build_options::RunMode,
@@ -58,22 +56,26 @@ use moonutil::{
     compiler_flags,
     cond_expr::OptLevel as BuildProfile,
     constants::{BLACKBOX_TEST_PATCH, MOONBITLANG_CORE, WHITEBOX_TEST_PATCH},
-    features::FeatureGate,
     package::SupportedTargetsDeclKind,
     project::{PackageDirs, ProjectManifest},
-    render::MooncDiagnostic,
     target::TargetBackend,
-    test_metadata::DiagnosticLevel,
     user_log::UserLog,
 };
 use tracing::{Level, info, instrument};
 
-use crate::build_flags::{BuildFlags, OutputStyle};
+use crate::build_flags::BuildFlags;
 
 pub mod action_identity;
 mod dry_run;
+mod execution;
 mod prebuild;
-pub use dry_run::{format_dry_run_command, write_dry_run, write_dry_run_all};
+#[cfg(test)]
+pub(crate) use dry_run::write_build_graph;
+pub(crate) use dry_run::{format_dry_run_command, write_dry_run};
+pub(crate) use execution::{
+    BuildConfig, JsonBuildOutput, execute_build, execute_build_json, execute_build_partial,
+    execute_test_build,
+};
 
 /// Synchronize dependencies and return resolved project data.
 /// This step does not acquire the target-directory lock.
@@ -534,11 +536,9 @@ pub(crate) fn plan_resolved_build_from_intent(
         artifact_paths: cx.artifact_paths.clone(),
     };
 
-    let db_path = cx.artifact_paths.target_layout().n2_db_path();
     let input = BuildInput {
         execution_plan,
         action_backends,
-        db_path,
     };
 
     info!("Build planning completed successfully");
@@ -562,11 +562,6 @@ pub fn plan_fmt(
         project_manifest,
         user_log,
     )?);
-    let layout = TargetLayout::from_fmt_resolve_output(
-        target_dir.to_path_buf(),
-        resolved,
-        BuildProfile::Debug,
-    );
     let action_backends = execution_plan
         .action_ids()
         .map(|action| (action, None))
@@ -574,7 +569,6 @@ pub fn plan_fmt(
     Ok(BuildInput {
         execution_plan,
         action_backends,
-        db_path: layout.n2_db_path(),
     })
 }
 
@@ -740,78 +734,6 @@ fn resolve_parallelism(jobs: Option<usize>) -> usize {
     jobs.unwrap_or_else(|| *DEFAULT_PARALLELISM)
 }
 
-#[derive(Clone)]
-pub struct BuildConfig {
-    /// The level of parallelism to use. If `None`, will use the number of
-    /// available CPU cores.
-    parallelism: Option<usize>,
-    /// The output style for errors and warnings
-    output_style: OutputStyle,
-    /// Render no-location diagnostics above this level
-    render_no_loc: DiagnosticLevel,
-    /// Maximum number of diagnostics to display after deduplication.
-    diagnostic_limit: Option<usize>,
-
-    /// Generate metadata file `packages.json`
-    pub generate_metadata: bool,
-
-    /// Explain and warnings in diagnostics
-    pub explain_errors: bool,
-
-    /// Ask n2 to explain rerun reasons
-    pub n2_explain: bool,
-
-    /// Verbose output for build progress and command echo
-    verbose: bool,
-    suppress_progress: bool,
-
-    /// The patch file to use
-    pub patch_file: Option<PathBuf>,
-}
-
-impl BuildConfig {
-    pub(crate) fn from_flags(
-        flags: &BuildFlags,
-        unstable_features: &FeatureGate,
-        verbose: bool,
-    ) -> Self {
-        BuildConfig {
-            parallelism: flags.jobs,
-            output_style: flags.output_style(),
-            render_no_loc: flags.render_no_loc,
-            diagnostic_limit: flags.diagnostic_limit,
-            generate_metadata: false,
-            explain_errors: false,
-            n2_explain: unstable_features.rr_n2_explain,
-            verbose,
-            suppress_progress: false,
-            patch_file: None,
-        }
-    }
-
-    pub(crate) fn with_suppressed_progress(mut self, suppress_progress: bool) -> Self {
-        self.suppress_progress = suppress_progress;
-        self
-    }
-}
-
-impl Default for BuildConfig {
-    fn default() -> Self {
-        Self {
-            parallelism: None,
-            output_style: OutputStyle::Raw,
-            render_no_loc: DiagnosticLevel::Error,
-            diagnostic_limit: None,
-            generate_metadata: false,
-            explain_errors: false,
-            n2_explain: false,
-            verbose: false,
-            suppress_progress: false,
-            patch_file: None,
-        }
-    }
-}
-
 /// A complete execution plan and the context needed to execute it.
 #[derive(Debug, Clone)]
 pub struct BuildInput {
@@ -820,24 +742,6 @@ pub struct BuildInput {
 
     /// Target Backend for each action. Shared actions have no single backend.
     action_backends: HashMap<ActionId, Option<TargetBackend>>,
-
-    /// The n2 database for the selected target directory.
-    db_path: PathBuf,
-}
-
-#[cfg(test)]
-impl BuildInput {
-    pub(crate) fn n2_graph_for_test(
-        &self,
-    ) -> Result<
-        (
-            n2::graph::Graph,
-            moonbuild_rupes_recta::execution_plan::CommandArgMap,
-        ),
-        moonbuild_rupes_recta::execution_plan::N2AdapterError,
-    > {
-        self.execution_plan.all_to_n2_graph()
-    }
 }
 
 impl BuildInput {
@@ -849,7 +753,6 @@ impl BuildInput {
         let Some(second) = inputs.next() else {
             return Ok(first);
         };
-        let db_path = first.db_path.clone();
 
         let mut execution_plan = ExecutionPlan::default();
         let mut action_backends = HashMap::new();
@@ -858,10 +761,6 @@ impl BuildInput {
             .chain(std::iter::once(second))
             .chain(inputs)
         {
-            anyhow::ensure!(
-                input.db_path == db_path,
-                "cannot compose build inputs with different target layouts"
-            );
             let existing_actions = execution_plan.action_ids().collect::<HashSet<_>>();
             let remapped = execution_plan.merge(&input.execution_plan)?;
             for (old, new) in input.execution_plan.action_ids().zip(remapped) {
@@ -878,745 +777,17 @@ impl BuildInput {
         Ok(Self {
             execution_plan: Rc::new(execution_plan),
             action_backends,
-            db_path,
         })
     }
-
-    fn into_n2_execution(
-        self,
-    ) -> Result<N2ExecutionInput, moonbuild_rupes_recta::execution_plan::N2AdapterError> {
-        let adapted = self
-            .execution_plan
-            .adapt_to_n2(self.execution_plan.action_ids())?;
-        let (graph, _, action_by_build) = adapted.into_parts_with_actions();
-        let backend_by_build = action_by_build
-            .into_iter()
-            .map(|(build, action)| (build, self.action_backends.get(&action).copied().flatten()))
-            .collect();
-        Ok(N2ExecutionInput {
-            graph,
-            db_path: self.db_path,
-            backend_by_build,
-        })
-    }
-}
-
-struct N2ExecutionInput {
-    graph: n2::graph::Graph,
-    db_path: PathBuf,
-    backend_by_build: HashMap<n2::graph::BuildId, Option<TargetBackend>>,
 }
 
 pub(crate) fn compose_build_inputs(inputs: Vec<BuildInput>) -> anyhow::Result<BuildInput> {
     BuildInput::compose(inputs)
 }
 
-struct CapturedBuildExecution {
-    n_tasks_executed: Option<usize>,
-    action_outputs: Vec<CapturedActionOutput>,
-}
-
-struct CapturedActionOutput {
-    target_backend: Option<TargetBackend>,
-    content: ResultCatcher,
-}
-
-impl CapturedBuildExecution {
-    fn successful(&self) -> bool {
-        self.n_tasks_executed.is_some()
-    }
-
-    fn diagnostic_sources<'a>(
-        &'a self,
-        build_metas: impl IntoIterator<Item = &'a BuildMeta>,
-    ) -> Vec<CapturedDiagnosticSource<'a>> {
-        let build_metas = build_metas.into_iter().collect::<Vec<_>>();
-        let sole_build_meta = match build_metas.as_slice() {
-            [build_meta] => Some(*build_meta),
-            _ => None,
-        };
-        self.action_outputs
-            .iter()
-            .map(|output| {
-                let build_meta = output
-                    .target_backend
-                    .and_then(|backend| {
-                        build_metas
-                            .iter()
-                            .find(|meta| meta.target_backend() == backend)
-                            .copied()
-                    })
-                    .or(sole_build_meta);
-                CapturedDiagnosticSource {
-                    diagnostics: &output.content,
-                    build_succeeded: self.successful(),
-                    build_meta,
-                }
-            })
-            .collect()
-    }
-}
-
-/// Execute a build plan.
-///
-/// Takes ownership of the build graph and executes the actual build tasks.
-/// Returns just the build result - callers should use the resolve data and
-/// artifacts from the planning phase for any metadata they need.
-///
-/// The caller must hold the target-directory lock. All executions in
-/// that directory share one n2 database, and n2 does not lock it internally.
-#[instrument(skip_all)]
-pub fn execute_build(
-    cfg: &BuildConfig,
-    input: BuildInput,
-    target_dir: &Path,
-    user_log: &UserLog,
-) -> anyhow::Result<N2RunStats> {
-    let execution = execute_build_capturing(cfg, input, target_dir)?;
-    Ok(finish_captured_build(cfg, &execution, None, user_log))
-}
-
-/// Structured output from one build execution for a command-level JSON
-/// renderer. The executor does not write diagnostics or summaries itself.
-pub struct JsonBuildOutput {
-    pub n_tasks_executed: Option<usize>,
-    pub n_errors: usize,
-    pub n_warnings: usize,
-    pub hidden_errors: usize,
-    pub hidden_warnings: usize,
-    pub diagnostics: Vec<JsonBuildDiagnostic>,
-    pub non_diagnostic_output: Vec<String>,
-}
-
-/// One compiler diagnostic and the backend of the n2 action that emitted it.
-/// The command layer remains responsible for projecting this into its JSON
-/// schema.
-pub struct JsonBuildDiagnostic {
-    pub target_backend: Option<TargetBackend>,
-    pub value: serde_json::Value,
-}
-
-impl JsonBuildOutput {
-    pub fn successful(&self) -> bool {
-        self.n_tasks_executed.is_some()
-    }
-}
-
-/// Execute a build while returning all Moonc diagnostics to the CLI seam.
-pub fn execute_build_json(
-    cfg: &BuildConfig,
-    input: BuildInput,
-    target_dir: &Path,
-) -> anyhow::Result<JsonBuildOutput> {
-    let execution = execute_build_capturing(cfg, input, target_dir)?;
-    // Keep the existing per-backend diagnostic-limit semantics while all
-    // backends execute in one n2 graph. Shared actions have no backend and are
-    // collected in their own group.
-    let mut sources_by_backend = BTreeMap::new();
-    for output in &execution.action_outputs {
-        sources_by_backend
-            .entry(output.target_backend)
-            .or_insert_with(Vec::new)
-            .push(CapturedDiagnosticSource {
-                diagnostics: &output.content,
-                build_succeeded: execution.successful(),
-                build_meta: None,
-            });
-    }
-
-    let mut n_errors = 0;
-    let mut n_warnings = 0;
-    let mut hidden_errors = 0;
-    let mut hidden_warnings = 0;
-    let mut diagnostics = Vec::new();
-    let mut non_diagnostic_output = Vec::new();
-    for (target_backend, sources) in sources_by_backend {
-        let collected = collect_json_diagnostics(&sources, cfg, true);
-        n_errors += collected.processed.n_errors;
-        n_warnings += collected.processed.n_warnings;
-        hidden_errors += collected.processed.hidden_errors;
-        hidden_warnings += collected.processed.hidden_warnings;
-        diagnostics.extend(collected.diagnostics.into_iter().map(|content| {
-            JsonBuildDiagnostic {
-                target_backend,
-                value: serde_json::from_str(&content)
-                    .expect("collected Moonc diagnostic should remain valid JSON"),
-            }
-        }));
-        non_diagnostic_output.extend(collected.non_diagnostic_output);
-    }
-
-    Ok(JsonBuildOutput {
-        n_tasks_executed: execution.n_tasks_executed,
-        n_errors,
-        n_warnings,
-        hidden_errors,
-        hidden_warnings,
-        diagnostics,
-        non_diagnostic_output,
-    })
-}
-
-/// Execute a test build.
-///
-/// Test builds may report diagnostics for generated drivers using their
-/// source-tree paths. The test build metadata lets the diagnostic processing
-/// stage resolve those paths through the target layout.
-pub fn execute_test_build(
-    cfg: &BuildConfig,
-    input: BuildInput,
-    target_dir: &Path,
-    build_metas: &[&BuildMeta],
-    user_log: &UserLog,
-) -> anyhow::Result<N2RunStats> {
-    let execution = execute_build_capturing(cfg, input, target_dir)?;
-    let sources = execution.diagnostic_sources(build_metas.iter().copied());
-    let processed = process_captured_diagnostics(&sources, cfg);
-    processed.warn_if_limited(user_log);
-    Ok(N2RunStats {
-        n_tasks_executed: execution.n_tasks_executed,
-        n_errors: processed.n_errors,
-        n_warnings: processed.n_warnings,
-    })
-}
-
-/// Callback on the [`n2::work::Work`] to be done for target artifacts.
-type WantFileFn<'b> = dyn for<'a> FnOnce(&'a mut n2::work::Work) -> anyhow::Result<()> + 'b;
-
-/// Partially execute a build graph, same as [`execute_build`] otherwise.
-///
-/// Pass `want_files` callback to determine which artifacts to build.
-///
-/// This function is primarily used for rebuilding tests after snapshot test
-/// promotion.
-#[instrument(skip_all)]
-pub fn execute_build_partial(
-    cfg: &BuildConfig,
-    input: BuildInput,
-    target_dir: &Path,
-    build_meta: Option<&BuildMeta>,
-    user_log: &UserLog,
-    want_files: Box<WantFileFn>,
-) -> anyhow::Result<N2RunStats> {
-    let N2ExecutionInput {
-        graph,
-        db_path,
-        backend_by_build,
-    } = input.into_n2_execution()?;
-    let execution = execute_n2_graph_capturing(
-        cfg,
-        graph,
-        db_path,
-        backend_by_build,
-        target_dir,
-        want_files,
-    )?;
-    Ok(finish_captured_build(cfg, &execution, build_meta, user_log))
-}
-
-fn execute_build_capturing(
-    cfg: &BuildConfig,
-    input: BuildInput,
-    target_dir: &Path,
-) -> anyhow::Result<CapturedBuildExecution> {
-    let N2ExecutionInput {
-        graph,
-        db_path,
-        backend_by_build,
-    } = input.into_n2_execution()?;
-    let start_nodes = graph.get_start_nodes();
-    execute_n2_graph_capturing(
-        cfg,
-        graph,
-        db_path,
-        backend_by_build,
-        target_dir,
-        Box::new(|work| {
-            // Want only the leaf output files, not all files including stdlib.
-            for file_id in start_nodes {
-                work.want_file(file_id)?;
-            }
-            Ok(())
-        }),
-    )
-}
-
-fn execute_n2_graph_capturing(
-    cfg: &BuildConfig,
-    mut build_graph: n2::graph::Graph,
-    db_path: PathBuf,
-    backend_by_build: HashMap<n2::graph::BuildId, Option<TargetBackend>>,
-    target_dir: &Path,
-    want_files: Box<WantFileFn>,
-) -> anyhow::Result<CapturedBuildExecution> {
-    // Ensure target directory exists
-    std::fs::create_dir_all(target_dir).context(format!(
-        "Failed to create target directory: '{}'",
-        target_dir.display()
-    ))?;
-
-    db_path
-        .parent()
-        .map(std::fs::create_dir_all)
-        .transpose()
-        .with_context(|| {
-            format!(
-                "Failed to create parent for build cache DB at {}",
-                db_path.display()
-            )
-        })?;
-
-    // Generate n2 state
-
-    let mut hashes = n2::graph::Hashes::default();
-    let n2_db = n2::db::open(&db_path, &mut build_graph, &mut hashes)
-        .with_context(|| format!("Failed to open build cache DB at {}", db_path.display()))?;
-
-    let parallelism = resolve_parallelism(cfg.parallelism);
-
-    let (captured_output_sender, captured_output_receiver) = mpsc::channel();
-    let mut prog_console: Box<dyn n2::progress::Progress> = create_progress_console(
-        Some(Box::new(move |build_id, output: &str| {
-            let target_backend = backend_by_build
-                .get(&build_id)
-                .copied()
-                .expect("every n2 build should retain its action backend");
-            let mut captured = ResultCatcher::default();
-            for line in output.split('\n').filter(|line| !line.is_empty()) {
-                captured.append_content(line, None);
-            }
-            captured_output_sender
-                .send(CapturedActionOutput {
-                    target_backend,
-                    content: captured,
-                })
-                .expect("captured output receiver should outlive n2 progress");
-        })),
-        cfg.verbose,
-        cfg.suppress_progress,
-    );
-    let mut work = n2::work::Work::new(
-        build_graph,
-        hashes,
-        n2_db,
-        &n2::work::Options {
-            failures_left: Some(10), // FIXME: This value is to match legacy, but might TBD
-            parallelism,
-            explain: cfg.n2_explain,
-            adopt: false,
-            dirty_on_output: true,
-        },
-        &mut *prog_console,
-        n2::smallmap::SmallMap::default(),
-    );
-    want_files(&mut work).context("Failed to determine the files to be built")?;
-
-    // The actual execution done by the n2 executor
-    let res = work.run().context("Failed to run n2 graph");
-    drop(work);
-    drop(prog_console); // Ensure the progress bar won't mess with diagnostic output
-    let res = res?;
-    let action_outputs = captured_output_receiver.into_iter().collect();
-
-    Ok(CapturedBuildExecution {
-        n_tasks_executed: res,
-        action_outputs,
-    })
-}
-
-fn finish_captured_build(
-    cfg: &BuildConfig,
-    execution: &CapturedBuildExecution,
-    build_meta: Option<&BuildMeta>,
-    user_log: &UserLog,
-) -> N2RunStats {
-    let sources = execution.diagnostic_sources(build_meta);
-    let processed = process_captured_diagnostics(&sources, cfg);
-    processed.warn_if_limited(user_log);
-    N2RunStats {
-        n_tasks_executed: execution.n_tasks_executed,
-        n_errors: processed.n_errors,
-        n_warnings: processed.n_warnings,
-    }
-}
-
-fn should_render_non_diagnostic_build_output(cfg: &BuildConfig, build_succeeded: bool) -> bool {
-    !(cfg.suppress_progress && build_succeeded)
-}
-
-struct CapturedDiagnosticSource<'a> {
-    diagnostics: &'a ResultCatcher,
-    build_succeeded: bool,
-    build_meta: Option<&'a BuildMeta>,
-}
-
-struct ProcessedDiagnostics {
-    n_errors: usize,
-    n_warnings: usize,
-    hidden_errors: usize,
-    hidden_warnings: usize,
-}
-
-struct CollectedJsonDiagnostics {
-    processed: ProcessedDiagnostics,
-    diagnostics: Vec<String>,
-    non_diagnostic_output: Vec<String>,
-}
-
-impl ProcessedDiagnostics {
-    fn warn_if_limited(&self, user_log: &UserLog) {
-        if self.hidden_errors != 0 || self.hidden_warnings != 0 {
-            user_log.warn(format!(
-                "diagnostic output limited by --diagnostic-limit: {} errors and {} warnings were not displayed.",
-                self.hidden_errors, self.hidden_warnings
-            ));
-        }
-    }
-}
-
-fn rewrite_captured_diagnostic(
-    content: &str,
-    cfg: &BuildConfig,
-    build_meta: Option<&BuildMeta>,
-) -> String {
-    let Some(meta) = build_meta else {
-        return content.to_owned();
-    };
-    let layout = meta.artifact_paths.target_layout();
-    let packages = &meta.resolve_output.pkg_dirs;
-    let backend = meta.target_backend();
-
-    if cfg.output_style.needs_moonc_json() {
-        let Ok(mut value) = serde_json::from_str::<serde_json::Value>(content) else {
-            return content.to_owned();
-        };
-        let mut changed = false;
-        let mut diagnostics = vec![&mut value];
-        while let Some(diagnostic) = diagnostics.pop() {
-            let Some(object) = diagnostic.as_object_mut() else {
-                continue;
-            };
-            if let Some(serde_json::Value::String(path)) = object.get_mut("path")
-                && let Some(physical) =
-                    layout.generated_test_driver_diagnostic_path(packages, Path::new(path), backend)
-            {
-                *path = physical.to_string_lossy().into_owned();
-                changed = true;
-            }
-            if let Some(serde_json::Value::Array(children)) = object.get_mut("children") {
-                diagnostics.extend(children.iter_mut());
-            }
-        }
-        return if changed {
-            serde_json::to_string(&value).expect("diagnostic JSON should serialize")
-        } else {
-            content.to_owned()
-        };
-    }
-
-    let Some(prefix_start) = content.find(GENERATED_TEST_DRIVER_PREFIX) else {
-        return content.to_owned();
-    };
-    let Some(extension_end) = content[prefix_start..].find(".mbt") else {
-        return content.to_owned();
-    };
-    let path_end = prefix_start + extension_end + ".mbt".len();
-    let Some(physical) = layout.generated_test_driver_diagnostic_path(
-        packages,
-        Path::new(&content[..path_end]),
-        backend,
-    ) else {
-        return content.to_owned();
-    };
-    format!("{}{}", physical.display(), &content[path_end..])
-}
-
-fn collect_json_diagnostics(
-    sources: &[CapturedDiagnosticSource<'_>],
-    cfg: &BuildConfig,
-    retain_suppressed_output: bool,
-) -> CollectedJsonDiagnostics {
-    let mut catcher = ResultCatcher::default();
-    for source in sources {
-        catcher.n_errors += source.diagnostics.n_errors;
-        catcher.n_warnings += source.diagnostics.n_warnings;
-    }
-
-    let mut by_file = BTreeMap::<String, BTreeSet<(MooncDiagnostic, String)>>::new();
-    let mut non_diagnostic_output = Vec::new();
-    for source in sources {
-        for content in &source.diagnostics.content_writer {
-            let content = rewrite_captured_diagnostic(content, cfg, source.build_meta);
-            match serde_json::from_str::<MooncDiagnostic>(&content) {
-                Ok(diagnostic) => {
-                    if diagnostic_is_generated_test_driver_warning(&diagnostic) {
-                        continue;
-                    }
-                    by_file
-                        .entry(diagnostic.path.clone())
-                        .or_default()
-                        .insert((diagnostic, content));
-                }
-                Err(_) => {
-                    if retain_suppressed_output
-                        || should_render_non_diagnostic_build_output(cfg, source.build_succeeded)
-                    {
-                        non_diagnostic_output.push(content);
-                    }
-                }
-            }
-        }
-    }
-
-    let mut diagnostics = Vec::new();
-    let (hidden_errors, hidden_warnings) = match cfg.diagnostic_limit {
-        None => {
-            for file_diagnostics in by_file.values() {
-                for (diagnostic, content) in file_diagnostics {
-                    diagnostics.push(content.clone());
-                    catcher.append_diag(diagnostic);
-                }
-            }
-            (0, 0)
-        }
-        Some(limit) => {
-            let mut displayed = 0;
-            let mut hidden_errors = 0;
-            let mut total_warnings = 0;
-            let mut displayed_warnings = 0;
-            let mut non_errors = Vec::new();
-
-            for file_diagnostics in by_file.values() {
-                for (diagnostic, content) in file_diagnostics {
-                    if diagnostic_is_error(diagnostic) {
-                        if displayed < limit {
-                            diagnostics.push(content.clone());
-                            catcher.append_diag(diagnostic);
-                            displayed += 1;
-                        } else {
-                            hidden_errors += 1;
-                        }
-                        continue;
-                    }
-
-                    if diagnostic_is_warning(diagnostic) {
-                        total_warnings += 1;
-                    }
-                    if displayed < limit {
-                        non_errors.push((diagnostic, content));
-                    }
-                }
-            }
-
-            if displayed < limit {
-                for (diagnostic, content) in non_errors {
-                    diagnostics.push(content.clone());
-                    catcher.append_diag(diagnostic);
-                    displayed += 1;
-                    if diagnostic_is_warning(diagnostic) {
-                        displayed_warnings += 1;
-                    }
-                    if displayed == limit {
-                        break;
-                    }
-                }
-            }
-
-            let hidden_warnings = total_warnings - displayed_warnings;
-            catcher.n_errors += hidden_errors;
-            catcher.n_warnings += hidden_warnings;
-            (hidden_errors, hidden_warnings)
-        }
-    };
-
-    CollectedJsonDiagnostics {
-        processed: ProcessedDiagnostics {
-            n_errors: catcher.n_errors,
-            n_warnings: catcher.n_warnings,
-            hidden_errors,
-            hidden_warnings,
-        },
-        diagnostics,
-        non_diagnostic_output,
-    }
-}
-
-fn process_captured_diagnostics(
-    sources: &[CapturedDiagnosticSource<'_>],
-    cfg: &BuildConfig,
-) -> ProcessedDiagnostics {
-    if cfg.output_style == OutputStyle::Json {
-        let collected = collect_json_diagnostics(sources, cfg, false);
-        for content in &collected.non_diagnostic_output {
-            eprintln!("{content}");
-        }
-        for content in &collected.diagnostics {
-            println!("{content}");
-        }
-        return collected.processed;
-    }
-
-    let mut catcher = ResultCatcher::default();
-    for source in sources {
-        catcher.n_errors += source.diagnostics.n_errors;
-        catcher.n_warnings += source.diagnostics.n_warnings;
-    }
-    let mut hidden_errors_total = 0;
-    let mut hidden_warnings_total = 0;
-    let captured = sources.iter().flat_map(|source| {
-        source
-            .diagnostics
-            .content_writer
-            .iter()
-            .map(move |content| {
-                (
-                    rewrite_captured_diagnostic(content, cfg, source.build_meta),
-                    source.build_succeeded,
-                )
-            })
-    });
-
-    match cfg.output_style {
-        OutputStyle::Json => unreachable!(),
-        OutputStyle::Fancy => {
-            let mut by_file = BTreeMap::<String, BTreeSet<MooncDiagnostic>>::new();
-            for (content, build_succeeded) in captured {
-                match serde_json::from_str::<moonutil::render::MooncDiagnostic>(&content) {
-                    Ok(d) => {
-                        if diagnostic_is_generated_test_driver_warning(&d) {
-                            continue;
-                        }
-                        by_file.entry(d.path.clone()).or_default().insert(d);
-                    }
-                    Err(_) => {
-                        // Non-diagnostics output, just print as-is
-                        // This could happen for installing binaries dependencies etc.
-                        if should_render_non_diagnostic_build_output(cfg, build_succeeded) {
-                            eprintln!("{content}");
-                        }
-                    }
-                };
-            }
-
-            let patch_file = cfg.patch_file.as_ref();
-            match cfg.diagnostic_limit {
-                None => {
-                    for file_diagnostics in by_file.values() {
-                        for diag in file_diagnostics {
-                            let kind = diag.render_diagnostics(
-                                n2::terminal::use_fancy(),
-                                patch_file,
-                                cfg.explain_errors,
-                                cfg.render_no_loc,
-                            );
-                            catcher.append_kind(kind);
-                        }
-                    }
-                }
-                Some(limit) => {
-                    let build_config = cfg;
-                    let mut displayed = 0;
-                    let mut hidden_errors = 0;
-                    let mut total_warnings = 0;
-                    let mut displayed_warnings = 0;
-                    let mut non_errors = Vec::new();
-
-                    for file_diagnostics in by_file.values() {
-                        for diag in file_diagnostics {
-                            if !diagnostic_is_renderable(diag, build_config) {
-                                continue;
-                            }
-
-                            if diagnostic_is_error(diag) {
-                                if displayed < limit {
-                                    let kind = diag.render_diagnostics(
-                                        n2::terminal::use_fancy(),
-                                        patch_file,
-                                        build_config.explain_errors,
-                                        build_config.render_no_loc,
-                                    );
-                                    catcher.append_kind(kind);
-                                    displayed += 1;
-                                } else {
-                                    hidden_errors += 1;
-                                }
-                                continue;
-                            }
-
-                            if diagnostic_is_warning(diag) {
-                                total_warnings += 1;
-                            }
-                            if displayed < limit {
-                                non_errors.push(diag);
-                            }
-                        }
-                    }
-
-                    if displayed < limit {
-                        for diag in non_errors {
-                            let kind = diag.render_diagnostics(
-                                n2::terminal::use_fancy(),
-                                patch_file,
-                                build_config.explain_errors,
-                                build_config.render_no_loc,
-                            );
-                            catcher.append_kind(kind);
-                            displayed += 1;
-                            if diagnostic_is_warning(diag) {
-                                displayed_warnings += 1;
-                            }
-                            if displayed == limit {
-                                break;
-                            }
-                        }
-                    }
-
-                    let hidden_warnings = total_warnings - displayed_warnings;
-                    hidden_errors_total += hidden_errors;
-                    hidden_warnings_total += hidden_warnings;
-                    catcher.n_errors += hidden_errors;
-                    catcher.n_warnings += hidden_warnings;
-                }
-            }
-        }
-        OutputStyle::Raw => {
-            for (content, _) in captured {
-                println!("{content}");
-            }
-        }
-    }
-    ProcessedDiagnostics {
-        n_errors: catcher.n_errors,
-        n_warnings: catcher.n_warnings,
-        hidden_errors: hidden_errors_total,
-        hidden_warnings: hidden_warnings_total,
-    }
-}
-
-fn diagnostic_is_error(diag: &MooncDiagnostic) -> bool {
-    diag.level == "error"
-}
-
-fn diagnostic_is_warning(diag: &MooncDiagnostic) -> bool {
-    matches!(diag.level.as_str(), "warn" | "warning")
-}
-
-fn diagnostic_is_generated_test_driver_warning(diag: &MooncDiagnostic) -> bool {
-    diagnostic_is_warning(diag) && diag.path.contains("__generated_driver_for_")
-}
-
-fn diagnostic_is_renderable(diag: &MooncDiagnostic, cfg: &BuildConfig) -> bool {
-    if !diag.path.is_empty() {
-        return true;
-    }
-
-    DiagnosticLevel::from_str(&diag.level, true).is_ok_and(|level| level >= cfg.render_no_loc)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use moonutil::render::{Loc, Position};
 
     #[cfg(unix)]
     #[test]
@@ -1635,119 +806,5 @@ mod tests {
         write_metadata_if_changed(&path, "second").unwrap();
         assert_ne!(path.metadata().unwrap().ino(), first_inode);
         assert_eq!(std::fs::read_to_string(path).unwrap(), "second");
-    }
-
-    fn diagnostic(path: &str, level: &str) -> MooncDiagnostic {
-        MooncDiagnostic {
-            path: path.to_string(),
-            loc: Loc {
-                start: Position { line: 1, col: 1 },
-                end: Position { line: 1, col: 2 },
-            },
-            level: level.to_string(),
-            message: String::new(),
-            error_code: 0,
-            children: Vec::new(),
-        }
-    }
-
-    #[test]
-    fn suppresses_generated_test_driver_warnings_only() {
-        assert!(diagnostic_is_generated_test_driver_warning(&diagnostic(
-            "./_build/wasm-gc/debug/test/lib/__generated_driver_for_internal_test.mbt",
-            "warning"
-        )));
-        assert!(!diagnostic_is_generated_test_driver_warning(&diagnostic(
-            "./_build/wasm-gc/debug/test/lib/__generated_driver_for_internal_test.mbt",
-            "error"
-        )));
-        assert!(!diagnostic_is_generated_test_driver_warning(&diagnostic(
-            "./lib/hello.mbt",
-            "warning"
-        )));
-    }
-
-    #[test]
-    fn generated_test_driver_errors_are_counted() {
-        let generated_driver_path =
-            "./_build/wasm-gc/debug/test/lib/__generated_driver_for_internal_test.mbt";
-        let warning = diagnostic(generated_driver_path, "warning");
-        let error = diagnostic(generated_driver_path, "error");
-        let mut catcher = ResultCatcher::default();
-        catcher.append_content(serde_json::to_string(&warning).unwrap(), None);
-        catcher.append_content(serde_json::to_string(&error).unwrap(), None);
-
-        let cfg = BuildConfig {
-            output_style: OutputStyle::Json,
-            ..Default::default()
-        };
-        let processed = process_captured_diagnostics(
-            &[CapturedDiagnosticSource {
-                diagnostics: &catcher,
-                build_succeeded: false,
-                build_meta: None,
-            }],
-            &cfg,
-        );
-
-        assert_eq!(processed.n_warnings, 0);
-        assert_eq!(processed.n_errors, 1);
-    }
-
-    #[test]
-    fn diagnostic_limit_is_shared_across_captured_build_errors() {
-        let mut dependency_error = diagnostic("./dependency.mbt", "error");
-        dependency_error
-            .children
-            .push(diagnostic("./dependency-detail.mbt", "error"));
-        let script_error = diagnostic("./script.mbt", "error");
-        let mut dependency = ResultCatcher::default();
-        dependency.append_content(serde_json::to_string(&dependency_error).unwrap(), None);
-        let mut script = ResultCatcher::default();
-        script.append_content(serde_json::to_string(&script_error).unwrap(), None);
-
-        let cfg = BuildConfig {
-            output_style: OutputStyle::Json,
-            diagnostic_limit: Some(1),
-            ..Default::default()
-        };
-        let processed = process_captured_diagnostics(
-            &[
-                CapturedDiagnosticSource {
-                    diagnostics: &dependency,
-                    build_succeeded: true,
-                    build_meta: None,
-                },
-                CapturedDiagnosticSource {
-                    diagnostics: &script,
-                    build_succeeded: false,
-                    build_meta: None,
-                },
-            ],
-            &cfg,
-        );
-
-        assert_eq!(processed.n_errors, 2);
-        assert_eq!(processed.n_warnings, 0);
-        assert_eq!(processed.hidden_errors, 1);
-        assert_eq!(processed.hidden_warnings, 0);
-    }
-
-    #[test]
-    fn structured_json_retains_successful_output_when_progress_is_suppressed() {
-        let mut catcher = ResultCatcher::default();
-        catcher.append_content("PREBUILD_SUCCESS", None);
-        let cfg = BuildConfig::default().with_suppressed_progress(true);
-        let sources = [CapturedDiagnosticSource {
-            diagnostics: &catcher,
-            build_succeeded: true,
-            build_meta: None,
-        }];
-
-        let collected = collect_json_diagnostics(&sources, &cfg, true);
-        assert_eq!(collected.non_diagnostic_output, ["PREBUILD_SUCCESS"]);
-
-        let rendered = collect_json_diagnostics(&sources, &cfg, false);
-        assert!(rendered.non_diagnostic_output.is_empty());
     }
 }
