+++ crates/moonbuild-debug/Cargo.toml
@@ -25,7 +25,5 @@ edition.workspace = true
 [dependencies]
 
 anyhow.workspace = true
-moonutil.workspace = true
-n2.workspace = true
 serde.workspace = true
 serde_json.workspace = true
