+++ crates/moonbuild/src/lib.rs
@@ -21,7 +21,6 @@
 pub mod bench;
 pub mod benchmark;
 pub mod doc_http;
-pub mod dry_run;
 pub mod entry;
 pub mod expect;
 pub mod new;
