+++ docs/dev/README.md
@@ -145,25 +145,23 @@ cargo install --path ./crates/moon --debug --offline
       binary
     - `src/cli/generate_test_driver.rs`: as the name suggests
   - `src/rr_build`: integration with the Rupes Recta build engine
+    - `execution/n2.rs`: n2 graph adaptation, incremental state, and execution
+    - `dry_run.rs`: renders commands and graph snapshots from Execution Plans
   - `tests/test_cases`: end-to-end tests organized into modules by purpose;
     `mod.rs` contains their shared imports and module registrations
 
 - `crates/moonbuild-rupes-recta`: the new build graph generation engine (now default)
-  - `src/build_lower`: lowers resolved modules to n2 build commands
+  - `src/build_lower`: lowers Build Plans to executor-neutral Execution Plans
   - `src/fmt.rs`: formatting support
   - `src/metadata.rs`: metadata generation for IDE/tooling
   - See `docs/dev/reference/compiler-cmd-ref.md` for compiler command reference
 
-- `crates/moonbuild`: the legacy build graph generation engine
-  - Being phased out in favor of `moonbuild-rupes-recta`
-  - `src/{check, gen, build, bundle, entry, runtest}`: generate
-    commands and n2 state according to `moon.mod.json` and `moon.pkg.json`
-  - `src/bundle.rs`: only for `moonbitlang/core`, not visible
-    to users
-  - `src/dry_run.rs`: prints commands without executing them,
-    mainly used by end-to-end tests.
-  - `src/expect.rs`: the implementation of expect tests in
-    `moon`
+- `crates/moonbuild`: support code retained from the former build engine
+  - `src/{entry, runtest, section_capture}`: test arguments, results, formatting,
+    and child-output capture
+  - `src/expect.rs`: expect and snapshot comparison and promotion
+  - `src/{bench, benchmark}`: benchmark generation and result handling
+  - `src/{new, upgrade, doc_http}`: project creation, self-upgrade, and documentation serving
 
 - `crates/mooncake`: package manager
   - `src/pkg/add`: `moon add`
@@ -186,7 +184,7 @@ cargo install --path ./crates/moon --debug --offline
 
 - `crates/moonrun`: runtime for executing WASM MoonBit programs
 
-- `crates/moonbuild-debug`: debugging utilities for dry-run printing and snapshotting
+- `crates/moonbuild-debug`: graph snapshot data and JSONL reading/writing used by tests
 
 ## Before PR
